+++ lib/features/plant/presentation/models/plant_registration_place.dart
@@ -1,0 +1,11 @@
+class PlantRegistrationPlace {
+  const PlantRegistrationPlace({
+    required this.id,
+    required this.name,
+    required this.imageAsset,
+  });
+
+  final String id;
+  final String name;
+  final String imageAsset;
+}
